pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = [
    "hatchling",
]

[project]
name = "allianceauth-georgeforge"
description = ""
readme = "README.md"
keywords = [
    "allianceauth",
]
license = { file = "LICENSE" }
authors = [  ]
requires-python = ">=3.10"
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]
dynamic = [
    "version",
]
dependencies = [
    "allianceauth<5,>=4",
    "django-eveuniverse<2,>=1",
]

[project.optional-dependencies]
discord = [
    "py-cord<3,>=2",
    "allianceauth-discordbot<5,>=4",
]
webhook = [
    "py-cord<3,>=2",
]

dev = [
    "django-stubs"
]

[tool.hatch.version]
path = "georgeforge/__init__.py"

[tool.hatch.build]
include = [ "/georgeforge" ]
